Move majority of LACP implementation out of valve.py and into standalone/stack.
