+++ faucet/valve_switch.py
@@ -61,6 +61,9 @@ def valve_switch_factory(logger, dp, pipeline, acl_manager):  # pylint: disable=
         'learn_jitter': dp.learn_jitter,
         'cache_update_guard_time': dp.cache_update_guard_time,
         'idle_dst': dp.idle_dst,
+        'dp_high_priority': dp.high_priority,
+        'dp_highest_priority': dp.highest_priority,
+        'faucet_dp_mac': dp.faucet_dp_mac,
     }
 
     if dp.stack_graph:
